requirements/base.txt
certifi==2018.4.16
Django==1.11
djangorestframework==3.8.2
pytz==2018.5

